WIP: Snowflake glue implementation
Refactor DMS template to be generic Incremental merge template
