+++ src/main/resources/templates/shared/snowflake/create-stage.ssp
@@ -13,11 +13,8 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 --%>
-#import(io.phdata.streamliner.configuration.Snowflake)
-<%@ val configuration: io.phdata.streamliner.configuration.Configuration %>
-#{
-    val destination = configuration.destination.asInstanceOf[Snowflake]
-}#
-CREATE STAGE IF NOT EXISTS ${destination.stagingDatabase.name}.${destination.stagingDatabase.schema}.${configuration.name}_stage
+<%@ val destination: io.phdata.streamliner.configuration.Snowflake %>
+<%@ val stage: String %>
+CREATE STAGE IF NOT EXISTS ${destination.stagingDatabase.name}.${destination.stagingDatabase.schema}.${stage}
     URL = '${destination.storagePath}'
     STORAGE_INTEGRATION = ${destination.storageIntegration};
