+++ src/main/resources/templates/shared/snowflake/drop-stage.ssp
@@ -13,9 +13,6 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 --%>
-#import(io.phdata.streamliner.configuration.Snowflake)
-<%@ val configuration: io.phdata.streamliner.configuration.Configuration %>
-#{
-    val destination = configuration.destination.asInstanceOf[Snowflake]
-}#
-CREATE SCHEMA IF NOT EXISTS ${destination.reportingDatabase.name}.${destination.reportingDatabase.schema};
+<%@ val database: io.phdata.streamliner.configuration.SnowflakeDatabase %>
+<%@ val stage: String %>
+DROP STAGE IF EXISTS ${database.name}.${database.schema}.${stage};
