+++ src/main/resources/templates/shared/snowflake/drop-task.ssp
@@ -13,10 +13,6 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 --%>
-#import(io.phdata.streamliner.configuration.Snowflake)
-<%@ val configuration: io.phdata.streamliner.configuration.Configuration %>
-<%@ val table: io.phdata.streamliner.configuration.TableDefinition %>
-#{
-    val destination = configuration.destination.asInstanceOf[Snowflake]
-}#
-DROP STREAM IF EXISTS ${destination.stagingDatabase.name}.${destination.stagingDatabase.schema}.${table.destinationName}_STREAM;
+<%@ val database: io.phdata.streamliner.configuration.SnowflakeDatabase %>
+<%@ val task: String %>
+DROP TASK IF EXISTS ${database.name}.${database.schema}.${task};
