+++ src/main/resources/templates/snowflake-incremental-merge/Makefile.ssp
@@ -104,6 +104,5 @@ clean:
     $(MAKE) drop-task
     $(MAKE) drop-snowpipe
     $(MAKE) drop-stream
-    $(MAKE) drop-stage
     $(MAKE) drop-report-table
     $(MAKE) drop-staging-table
