+++ src/main/resources/templates/snowflake-incremental-merge/alter-task-resume.sql.ssp
