+++ src/main/resources/templates/snowflake-incremental-merge/alter-task-suspend.sql.ssp
