+++ src/main/resources/templates/snowflake-incremental-merge/copy-bulk-load-into-stg.sql.ssp
@@ -22,11 +22,13 @@ limitations under the License.
     val destColumnStr = table.columns.map(column => s"${column.destinationName}").mkString(", ")
     val sourceColumnStr = table.columns.map(column => s"$$1:${column.sourceName}::${io.phdata.streamliner.util.TemplateFunction.mapDataType(column, typeMapping, "SNOWFLAKE")}").mkString(", ")
 }#
+
 USE DATABASE ${destination.stagingDatabase.name};
 
 USE SCHEMA ${destination.stagingDatabase.schema};
 
-COPY INTO departments_stg (${destColumnStr}, dms_operation, dms_ts)
+COPY INTO ${table.destinationName}_stg (${destColumnStr}, dms_operation, dms_ts)
 FROM ( SELECT ${sourceColumnStr}, $1:Op::STRING, $1:ts::TIMESTAMP_LTZ
-    FROM @MYSQL_DMS_EXAMPLE_stage/departments)
+        FROM @${configuration.name}_stage/${table.sourceName})
+PATTERN = '^.*\/LOAD[0-9]{8}\.parquet$'
 FILE_FORMAT = ( TYPE = PARQUET);
