+++ src/main/resources/templates/snowflake-incremental-merge/create-snowpipe.sql.ssp
@@ -1,0 +1,35 @@
+<%--
+Copyright 2018 phData Inc.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+--%>
+#import(io.phdata.streamliner.configuration.Snowflake)
+<%@ val configuration: io.phdata.streamliner.configuration.Configuration %>
+<%@ val table: io.phdata.streamliner.configuration.TableDefinition %>
+<%@ val typeMapping: Map[String, Map[String, String]]%>
+#{
+    val destination = configuration.destination.asInstanceOf[Snowflake]
+    val destColumnStr = table.columns.map(column => s"${column.destinationName}").mkString(", ")
+    val sourceColumnStr = table.columns.map(column => s"$$1:${column.sourceName}::${io.phdata.streamliner.util.TemplateFunction.mapDataType(column, typeMapping, "SNOWFLAKE")}").mkString(", ")
+}#
+USE DATABASE ${destination.stagingDatabase.name};
+
+USE SCHEMA ${destination.stagingDatabase.schema};
+
+CREATE PIPE IF NOT EXISTS ${table.destinationName}_pipe
+    AUTO_INGEST = true
+    AS
+        COPY INTO ${table.destinationName}_stg (${destColumnStr}, dms_operation, dms_ts)
+            FROM ( SELECT ${sourceColumnStr}, $1:Op::STRING, $1:ts::TIMESTAMP_LTZ
+                   FROM @${configuration.name}_stage/${table.sourceName})
+        FILE_FORMAT = ( TYPE = PARQUET);
