+++ src/main/resources/templates/snowflake-incremental-merge/create-staging-schema.sql.ssp
@@ -15,8 +15,7 @@ limitations under the License.
 --%>
 #import(io.phdata.streamliner.configuration.Snowflake)
 <%@ val configuration: io.phdata.streamliner.configuration.Configuration %>
-<%@ val table: io.phdata.streamliner.configuration.TableDefinition %>
 #{
     val destination = configuration.destination.asInstanceOf[Snowflake]
 }#
-DROP PIPE IF EXISTS ${destination.stagingDatabase.name}.${destination.stagingDatabase.schema}.${table.destinationName}_pipe;
+<% render("../shared/snowflake/create-schema.ssp", Map("database" -> destination.stagingDatabase)) %>
