+++ src/main/resources/templates/snowflake-incremental-merge/create-staging-table.sql.ssp
