+++ src/main/resources/templates/snowflake-incremental-merge/drop-report-table.sql.ssp
@@ -1,0 +1,22 @@
+<%--
+Copyright 2018 phData Inc.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+--%>
+#import(io.phdata.streamliner.configuration.Snowflake)
+<%@ val configuration: io.phdata.streamliner.configuration.Configuration %>
+<%@ val table: io.phdata.streamliner.configuration.TableDefinition %>
+#{
+    val destination = configuration.destination.asInstanceOf[Snowflake]
+}#
+<% render("../shared/snowflake/drop-table.ssp", Map("database" -> destination.reportingDatabase, "table" -> table.destinationName)) %>
