+++ src/main/resources/templates/snowflake-snowpipe-append/Makefile.ssp
@@ -1,0 +1,63 @@
+%--
+Copyright 2018 phData Inc.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+--%>
+#import(io.phdata.streamliner.configuration.Snowflake)
+<%@ val configuration: io.phdata.streamliner.configuration.Configuration %>
+<%@ val table: io.phdata.streamliner.configuration.TableDefinition %>
+#{
+    val destination = configuration.destination.asInstanceOf[Snowflake]
+}#
+snowsql-cmd=${destination.snowSqlCommand} -f
+
+copy-into: copy-into.sql
+    $(snowsql-cmd) copy-into.sql
+
+create-schema: create-schema.sql
+    $(snowsql-cmd) create-schema.sql
+
+create-snowpipe: create-snowpipe.sql
+    $(snowsql-cmd) create-snowpipe.sql
+
+create-stage: create-stage.sql
+    $(snowsql-cmd) create-stage.sql
+
+create-table: create-table.sql
+    $(snowsql-cmd) create-table.sql
+
+drop-schema: drop-schema.sql
+    $(snowsql-cmd) drop-schema.sql
+
+drop-snowpipe: drop-snowpipe.sql
+    $(snowsql-cmd) drop-snowpipe.sql
+
+drop-stage: drop-stage.sql
+    $(snowsql-cmd) drop-stage.sql
+
+drop-table: drop-table.sql
+    $(snowsql-cmd) drop-table.sql
+
+first-run:
+    $(MAKE) create-schema
+    $(MAKE) create-stage
+    $(MAKE) create table
+    $(MAKE) copy-into
+    $(MAKE) create-snowpipe
+
+clean:
+    $(MAKE) drop-schema
+    $(MAKE) drop table
+    $(MAKE) drop-snowpipe
+
+
