+++ src/main/resources/templates/snowflake-snowpipe-append/create-table.sql.ssp
@@ -19,16 +19,5 @@ limitations under the License.
 <%@ val typeMapping: Map[String, Map[String, String]]%>
 #{
     val destination = configuration.destination.asInstanceOf[Snowflake]
-    val destColumnStr = table.columns.map(column => s"${column.destinationName}").mkString(", ")
-    val sourceColumnStr = table.columns.map(column => s"$$1:${column.sourceName}::${io.phdata.streamliner.util.TemplateFunction.mapDataType(column, typeMapping, "SNOWFLAKE")}").mkString(", ")
 }#
-
-USE DATABASE ${destination.stagingDatabase.name};
-
-USE SCHEMA ${destination.stagingDatabase.schema};
-
-COPY INTO departments_stg (${destColumnStr}, dms_operation, dms_ts)
-FROM ( SELECT ${sourceColumnStr}, $1:Op::STRING, $1:ts::TIMESTAMP_LTZ
-        FROM @MYSQL_DMS_EXAMPLE_stage/departments)
-PATTERN = '^.*\/LOAD[0-9]{8}\.parquet$'
-FILE_FORMAT = ( TYPE = PARQUET);
+<% render("../shared/snowflake/create-table.ssp", Map("configuration" -> configuration, "table" -> table, "typeMapping" -> typeMapping))%>
