+++ src/main/scala/io/phdata/streamliner/configuration/JDBCParser.scala
@@ -46,38 +46,12 @@ object JDBCParser {
       tables = Some(tables)
     )
 
-    checkConfiguration(enhancedConfiguration)
     enhancedConfiguration
   }
 
-  private def checkConfiguration(configuration: Configuration): Unit = {
-    for (table <- configuration.tables.get) {
-      if (configuration.pipeline.equalsIgnoreCase("INCREMENTAL-WITH-KUDU")) {
-        checkPrimaryKeys(table)
-        checkCheckColumn(table)
-      } else if (configuration.pipeline.equalsIgnoreCase("KUDU-TABLE-DLL") || configuration.pipeline
-        .equalsIgnoreCase("SNOWFLAKE-DMS-CDC")) {
-        checkPrimaryKeys(table)
-      }
-    }
-  }
-
-  private def checkPrimaryKeys(table: TableDefinition): Unit = {
-    if (table.primaryKeys.isEmpty) {
-      logger.warn(s"No primary keys are defined for table: ${table.sourceName}")
-    }
-  }
-
-  private def checkCheckColumn(table: TableDefinition): Unit = {
-    if (table.checkColumn.isEmpty) {
-      logger.warn(
-        s"No check column is defined for table: ${table.sourceName}, sqoop incremental import will fail for this table unless the checkColumn is added manually")
-    }
-  }
-
   private def mapTableDefinition(
-                                  table: SchemaCrawlerTable,
-                                  metadataOpt: Option[Map[String, String]] = None): TableDefinition = {
+      table: SchemaCrawlerTable,
+      metadataOpt: Option[Map[String, String]] = None): TableDefinition = {
     TableDefinition(
       sourceName = table.getName,
       destinationName = TemplateFunction.cleanse(table.getName),
@@ -89,10 +63,8 @@ object JDBCParser {
   }
 
   private def enhanceTableDefinition(
-                                      table: TableDefinition,
-                                      userDefined: Table): TableDefinition = {
-    checkNumberOfMappers(userDefined)
-
+      table: TableDefinition,
+      userDefined: Table): TableDefinition = {
     val enhancedMetadata = table.metadata match {
       case Some(metadata) =>
         userDefined.metadata match {
@@ -120,17 +92,4 @@ object JDBCParser {
       precision = Option(column.getSize),
       scale = Option(column.getDecimalDigits)
     )
-
-  private def checkNumberOfMappers(table: Table) = {
-    table.numberOfMappers match {
-      case Some(numberOfMappers) =>
-        if (numberOfMappers > 1) {
-          if (table.splitByColumn.isEmpty) {
-            throw new RuntimeException(
-              s"Table: $table, has number of mappers greater than 1 with no splitByColumn defined Sqoop import will fail for this table")
-          }
-        }
-      case None => // no-op
-    }
-  }
 }
